video-translation-webhook-backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.1
pycryptodome==3.20.0
gevent==24.2.1
gevent-websocket==0.10.1 